Propose reference to the open web.
When I watched/listened to this, I was really drawn through the narrative, but felt like it was missing something to tie me back to what it was about, and kind of "close" the section about the diverse group of people, their freedoms etc. I've added a simple "title slide"/statement that goes along/contrasts with the Facebook/Instagram ones, to try to capture that.